base::IReadOnlyCollection<int, bsp::IKey *> &DI_KeyCollection(); bsp::IKeyScanner &DI_KeyScanner();

+++ include/bsp-interface/key/IKeyScanner.h
@@ -17,20 +17,20 @@ namespace bsp
 		/// @note 在再次调用 ScanKeys 之前，本函数的返回值会保持一致。
 		/// @param key_index
 		/// @return 指定的按键有按键按下动作则返回 true，否则返回 false。
-		virtual bool HasKeyDownEvent(uint16_t key_index) = 0;
+		virtual bool HasKeyDownEvent(int key_index) = 0;
 
 		/// @brief 检测指定的按键是否有抬起动作事件。
 		/// @note 所谓抬起动作，就是从按下变成松开的这个状态变化。本事件只有在这个边沿会触发一次。
 		/// @note 在再次调用 ScanKeys 之前，本函数的返回值会保持一致。
 		/// @param key_index
 		/// @return 指定的按键有按键抬起动作则返回 true，否则返回 false。
-		virtual bool HasKeyUpEvent(uint16_t key_index) = 0;
+		virtual bool HasKeyUpEvent(int key_index) = 0;
 
 		/// @brief 检测指定的按键是否有被按压着事件。
 		/// @note 所谓被按压着事件，就是最近一次调用 ScanKeys 得到的按键状态。即当前扫描的结果
 		/// 是按键正被按着。
 		/// @param key_index
 		/// @return 指定的按键正被按着则返回 true，否则返回 false。
-		virtual bool HasKeyPressedEvent(uint16_t key_index) = 0;
+		virtual bool HasKeyPressedEvent(int key_index) = 0;
 	};
 }
